chore: move toast to TriggerSyncButton component (#85)
Right now the toast logic is written inside the sample app.

This is step 1 of moving the entire config setting into nextjs package.

We should consider making the toast a primitive, but it's tricky because
it depends on its parent container. We should probably have a
`NotificationProvider` that manages all of these, but punting for now.