feat: enable customers to create custom properties in the sync config (#105)
Depends on #104

Adds a simple button to allow customers to create custom properties for
the purposes of their SFDC integration. Note that these are
application-side custom properties, not Salesforce custom fields.

For now, this is a fairly naive implementation that assumes that
customer-defined properties aren't consumed by the application outside
of the Salesforce syncs.

**Caveats**
1. Custom properties cannot be renamed
2. Custom properties cannot be reused across syncs, even for the same
sObjects (same as field mappings). We're planning to refactor how syncs
are modeled to enable field mappings, custom properties, etc. to be
shared across syncs.

TO DO (follow-up):
- [ ] Actually write ingested Salesforce data to the custom properties
- [ ] Validate custom field when creating and display errors
- [ ] Fix annoying UI issue where the new custom property input lingers
after the field mapping has been updated
- [ ] Only allow creating custom properties if enabled for that customer
by the developer
- [ ] Update SG default theme to include form/input styles (FYI
@asdfryan - I held off on this until you're back from OOO)