chore: add customPropertiesColumn to Postgres destination config (#107)
Add a new config property to indicate where customer-defined properties
should be written to in a Postgres destination. For now, this must be a
`jsonb` column on the same table that we're writing Salesforce objects
to.

Includes a migration to add a custom properties column to the
`salesforce_contacts` table in the sample app.