feat: add customPropertiesEnabled setting to sync configs (#150)
Adds a separate config property to the `SyncConfig` that governs whether
or not a customer is able to add custom fields to a field mapping. Note
that for postgres integrations, a `customPropertiesColumn` is also
required.

Also fixes an issue where `destination.schema.fields` is out of sync
with the `defaultFieldMapping`, so that fields specified in the latter
but not the former do not appear in the customer's field mapping.