fix: bump keepalive timeout to 61 seconds (#858)
I suspect that the 504s are for long running requests. Since we set the
keepalive timeout on the ELBs to 4 seconds to keep under the node
default of 5 seconds, its probable that the ELB is now timing out the
request if it takes longer than the 4 seconds.

After this is deployed, I will bump the ELB idle timeout back up to the
default 60 sec.