docs: add default admin credentials to quickstart docs (#1022)
<!-- Please title your PR using conventional commits
(https://www.conventionalcommits.org/en/v1.0.0/): -->

When following the quickstart documentation, you're instructed to go to
the management UI after `docker-compose up`, but the default admin
credentials aren't listed anywhere that I could see. This adds knowledge
of the defaults, as well as information on how to modify the defaults if
you wish.