fix: marketo fields (#1681)
Apparently for Marketo, when you submit a form, it expects the field
names to be a `Lead` field name (which are in camelCase), but the Form
Fields are in PascalCase. However, there seems to be no mapping between
the Form Field to the Lead Field.

E.g. if you try to submit a form with `FirstName` it will fail, but
`firstName` will work.

This is not a perfect solution, but I can't think of anything better.