chore: refactor svix away from getServerSideProps (#1716)
Context: we are calling a Svix call on every `getServerSideProps`, even
on pages that don't need it. This is adding ~200-300ms of latency to
each page load.

I've refactored it to move the Svix call to just that tab only. This
requires moving the Svix API key to `NEXT_PUBLIC_...`